Clear interrupts
You have to clear interrupts in an interrupt handler, otherwise the interrupt controller will keep calling the handler in a tight loop.

On my board, as of this commit, the monitor syncs to the output and I see a yellow-and-red digital snow pattern, which suggests that the blue channel is being generated correctly; it also responds to the first line of EMPTY_SCANLINE_TMDS being changed.

When testing with the LED at reduced frequencies, I've observed that the wait_for_load calls on lanes 1 and 2 seem to hang, but lane 0 seems to be correct. However, this doesn't seem to be a problem at full speed.