Introduce scanlist
This patch creates a virtual machine for rendering TMDS-encoded scanlines from a display list. The display is an improved version of the SMPTE colorbars.

Only solid color blocks are currently implemented, but the virtual machine is extensible and the next step is to convert low bit depth line buffers into full TMDS scanlines.