Display list for dynamically generating scenes (#12)
This patch puts a rendering pipeline into place, with a solid architecture though currently somewhat limited functionality. There are a number of parts to it.

The display list is split into two parts. The renderlist part renders content into a reduced bit depth buffer. The scanlist part takes those reduced bit depth buffers and encodes into TMDS. Both renderers are implemented as direct threaded virtual machines, with the individual operations written in assembly for performance.

The scanlist rendering happens on the video interrupt and is intended to run in less than one scanline time. The renderlist rendering is striped across both cores, preferentially running in core1 main, but if that is busy, it runs in an interrupt in core0.

The application rendering happens in core0 main. It acquires display list builders (which in general means waiting for vsync), builds the display list, then sends them to the system, which swaps them each frame if available. If main takes longer than a single frame time, then the previous frame is repeated.

Text is rendered to a 1bpp buffer, using a fixed palette.

There are some rough edges. Naming and organization could be improved. There is a cavalier approach to safety. Obviously more functionality in the renderers can be built. But this commit represents a significant advance.

---------

Co-authored-by: Zachary Kohnen <me@dusterthefirst.com>