Render tiles (#14)
A very simple demo with one tile, panning smoothly.

There's also a bit of reorganization with the demo moved to its own file. It's written so it can dynamically switch between different demos.

Also fixes an issue where the jump in the renderlist was set as a delta, but interpreted as an offset from the beginning. That works for one or two stripes, not more.