Go back to hand twiddling the bits
It looks like my handwritten code has a ~10x speedup on whatever bitvec was able to produce. This seems to be coming from my optimization on the center 30 bits of the word, not having to look at the neighboring bytes at all.

Since this is a microcontroller, and does not have any vector instructions (that I know of) I do not think there are many more speed savings to be had, at least in rust land.