Demo: Conway's Game of Life (#17)
* Resizable board

The board does not pack the rows densely yet. If the width is not a multiple of 32, there can be wasted space off the end of each row

* Loads worlds from a subset of Life_1.05

I have implemented a subset of Life_1.05, mainly #P X Y which lets the universe file place itself all over the board.

* Handwritten, but probably still naive implementation of new state calculations

I attempted to use `bitvec` as it touts itself as "as fast as you could write by hand", but it looks like my handwritten code has a ~10x speedup on whatever `bitvec` was able to produce. This seems to be coming from my optimization on the center 30 bits of the word not having to look at the neighboring bytes at all.

Since this is a microcontroller, and does not have any vector instructions (that I know of) I do not think there are many more speed savings to be had, at least in rust land.