Make interrupt timing consistent
Always end a line with a FIFO's worth full of raw or TMDS data. That causes the interrupt to happen very near the end of the scanline. Without this fix, the intrrupt happens at the beginning of the sync lines, and the inconsistent timing can cause problems, in particular the safety margin is inadequate at 32bpp.

See https://forums.raspberrypi.com/viewtopic.php?t=387507 for more analysis.