Port to RP2350 (#20)
* Port to RP2350

WIP, does solid color output, lots of unused warnings as demo isn't wired up yet.

This version may be useful for repro of DMA bandwidth limitations.

* Make interrupt timing consistent

Always end a line with a FIFO's worth full of raw or TMDS data. That causes the interrupt to happen very near the end of the scanline. Without this fix, the intrrupt happens at the beginning of the sync lines, and the inconsistent timing can cause problems, in particular the safety margin is inadequate at 32bpp.

See https://forums.raspberrypi.com/viewtopic.php?t=387507 for more analysis.

* Get rendering working on RP2350

WIP

In a bit of a frustrating state. It outputs vertical colorbars, but is very fragile, many perturbations cause it to fail.

* Switch cores 0 and 1

Run the DVI on core 0 and the app on core 1. This is more robust, and gets the colorbar example running. It is extremely unclear to me why.

* Working text rendering

Hook up new 1bpp palette and also make text rendering example work. Still on "demo2" and with some old stuff remaining, but should be able to clean that up shortly.

* Run complete demo

With this patch, the entire demo is ported over, including 4bpp palettes for the tile demo.

Note that the conway demo allocates the cell state on the stack, so requires rather a huge stack to run successfully.

* Minor cleanups

Try to reduce warnings. Delete things that are not needed.

There are some warnings from the use of static mut. Probably we should do the UnsafeCell thing, but it is more noise (especially with needing to implement Sync).