Move video out to core 1
The intent is to have video output running on core 1, so that the app can run on core 0. This wasn't working before because we were unmasking the DMA interrupt on core 0. The attempt to spawn core1_main to core 1 caused failures because the interrupt was then enabled on both cores.

This patch does DMA setup in correct order, so running both the interrupt and core1_main on core 1 actually works now.