New interrupt scheme (#21)
* Fix half-word increments

The asm syntax for post-increment in str is confusing, it doesn't match stmia.

* Start new rendering system

DMA interrupt fires twice per line. Just solid color for now.

* Hook up video lines

Use line buffers. Trivial content for now, but the structure is in place.

Some refactoring, largely so client uses safe methods.

* Hook up display list

Run the display list in core1_main. This is very simplistic from a scheduling point of view, it just runs one thread, and robustness logic is lacking if there's a miss, but it does run the demo.

* Move video out to core 1

The intent is to have video output running on core 1, so that the app can run on core 0. This wasn't working before because we were unmasking the DMA interrupt on core 0. The attempt to spawn core1_main to core 1 caused failures because the interrupt was then enabled on both cores.

This patch does DMA setup in correct order, so running both the interrupt and core1_main on core 1 actually works now.

* Make HSTX pinouts easier to configure

Introduce new DviPinout struct for representing board-specific HSTX pinouts (and related enums); pass this into `setup_hstx`.

This supports the Lilygo T-Pico-2 board without having to edit the code. It's also a step towards structuring the project for real board configurability.