Bring in rtic
Move the core0 main to an RTIC app. This is a minimal change that doesn't affect functionality, but should get things in place for USB host support through cotton-usb.

It brings in some git dependencies because USB host support is missing from rp235x-hal. Hopefully that will land and we can go back to released crates.