Working USB loopback signaling
This commit demonstrates working TX and RX signaling, configured as loopback. It displays the received bytes on the console.

A few changes will be necessary to actually use this code. The TX and RX programs require a total of 33 instruction slots, which don't quite fit. In this commit, the final J state is omitted, and both can run at the same time, but that is not spec compliant.

There are two alternate versions of TX, which do bit stuffing and NRZI, but require a 2-bit encoding of the input (a 256 entry LUT would do the trick). The code golfed one is only 8 instructions, so the juggling would not be necessary.

Some notes about clocks. This code is set up for a 126 MHz sysclk, which is sensible for 480p. That will incur some jitter on output but because the ratio is a multiple of 1/256 (it is 2.625) the frequency is accurate. It may be that the USB hub regenerates a clean clock (to be validated by measurement).

Clock accuracy for RX seems much less critical, as there is a low-to-high edge every 14 cycles. This RX program is tuned for 10 PIO instructions per bit but can possibly be tuned.